+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+* Test case that hard-resets the device and measures the time it takes to re-enumerate.
 * `DummyUsbBus` without functionality to allow examples that actually compile (but not run).
 * Extended `UsbRev` enum with variants for USB 1.0 and 1.1.